переработал LayerException на HttpResponseException Добавил возможность добавлять внутреннее исключение и логгировать его
