+++ src/Infrastructure/Middleware/HttpResponseExceptionMiddleware.cs
@@ -1,6 +1,8 @@
 ﻿using Infrastructure.Common.Exceptions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,13 +11,16 @@ using System.Threading.Tasks;
 
 namespace Infrastructure.Middleware
 {
-    public class LayerExceptionMiddleware
+    public class HttpResponseExceptionMiddleware
     {
         private readonly RequestDelegate _next;
 
-        public LayerExceptionMiddleware(RequestDelegate next)
+        private readonly ILogger<HttpResponseExceptionMiddleware> _logger;
+
+        public HttpResponseExceptionMiddleware(RequestDelegate next, ILogger<HttpResponseExceptionMiddleware> logger)
         {
             _next = next;
+            _logger = logger;
         }
 
         public async Task InvokeAsync(HttpContext context)
@@ -24,23 +29,27 @@ namespace Infrastructure.Middleware
             {
                 await _next(context);
             }
-            catch (LayerException ex)
+            catch (HttpResponseException ex)
             {
-
                 context.Response.Clear();
                 context.Response.StatusCode = (int)ex.StatusCode;
                 context.Response.ContentType = ex.ContentType;
                 await context.Response.WriteAsJsonAsync(ex.Error);
+
+                if (ex.InnerException != null)
+                {
+                    _logger.LogError(ex.InnerException, $"{ex.Message} because of {ex.InnerException.Message}");
+                }
                 return;
             }
         }
     }
 
-    public static class LayerExceptionMiddlewareExtensions
+    public static class HttpResponseExceptionMiddlewareExtensions
     {
-        public static IApplicationBuilder UseLayerExceptionMiddleware(this IApplicationBuilder builder)
+        public static IApplicationBuilder UseHttpResponseExceptionMiddleware(this IApplicationBuilder builder)
         {
-            return builder.UseMiddleware<LayerExceptionMiddleware>();
+            return builder.UseMiddleware<HttpResponseExceptionMiddleware>();
         }
     }
 }
